requirements.txt
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-stdimage==5.3.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
